skin_deploy/requirements.txt
torch
torchvision
fastapi
uvicorn
pillow
numpy

